Refactor headers in panda/board/stm32f4/*.h
